+++ CHANGELOG.md
@@ -97,6 +97,7 @@ Consider migrating to `stream-chat-android-ui-components` or `stream-chat-androi
 - Added support for Giphy command.
 - Added message pinning to the list of message options
 - Added pinned message UI
+- Added an option to flag a message to the message options overlay. 
 
 ### ⚠️ Changed
 - Changed the way focus state works for focused messages.